+++ crates/re_viewer/src/app.rs
@@ -935,12 +935,12 @@ fn file_menu(ui: &mut egui::Ui, app: &mut App, frame: &mut eframe::Frame) {
                     ui.spinner();
                 });
                 ui.horizontal(|ui| {
-                    let _ = ui.button("Save Time Selection…");
+                    let _ = ui.button("Save Loop Selection…");
                     ui.spinner();
                 });
             });
         } else {
-            let (clicked, time_selection) = ui
+            let (clicked, loop_selection) = ui
                 .add_enabled_ui(!app.log_db().is_empty(), |ui| {
                     if ui
                         .button("Save…")
@@ -950,33 +950,33 @@ fn file_menu(ui: &mut egui::Ui, app: &mut App, frame: &mut eframe::Frame) {
                         return (true, None);
                     }
 
-                    // We need to know the time selection _before_ we can even display the
+                    // We need to know the loop selection _before_ we can even display the
                     // button, as this will determine wether its grayed out or not!
                     // TODO(cmc): In practice the loop (green) selection is always there
                     // at the moment so...
-                    let time_selection = app
+                    let loop_selection = app
                         .state
                         .recording_configs
                         .get(&app.state.selected_rec_id)
-                        // is there an active time selection?
+                        // is there an active loop selection?
                         .and_then(|rec_cfg| {
                             rec_cfg
                                 .time_ctrl
-                                .time_selection()
+                                .loop_selection()
                                 .map(|q| (*rec_cfg.time_ctrl.timeline(), q))
                         });
 
                     if ui
                         .add_enabled(
-                            time_selection.is_some(),
-                            egui::Button::new("Save time selection…"),
+                            loop_selection.is_some(),
+                            egui::Button::new("Save loop selection…"),
                         )
                         .on_hover_text(
-                            "Save data for the current time selection to a Rerun data file (.rrd)",
+                            "Save data for the current loop selection to a Rerun data file (.rrd)",
                         )
                         .clicked()
                     {
-                        return (true, time_selection);
+                        return (true, loop_selection);
                     }
 
                     (false, None)
@@ -988,7 +988,7 @@ fn file_menu(ui: &mut egui::Ui, app: &mut App, frame: &mut eframe::Frame) {
                 // the DB isn't empty: let's spawn a new one.
 
                 if let Some(path) = rfd::FileDialog::new().set_file_name("data.rrd").save_file() {
-                    let f = save_database_to_file(app, path, time_selection);
+                    let f = save_database_to_file(app, path, loop_selection);
                     if let Err(err) = app.spawn_threaded_promise(FILE_SAVER_PROMISE, f) {
                         // NOTE: Shouldn't even be possible as the "Save" button is already
                         // grayed out at this point... better safe than sorry though.
@@ -1158,6 +1158,7 @@ fn save_database_to_file(
             .chronological_log_messages()
             .cloned()
             .collect::<Vec<_>>(),
+
         // Query path: time to filter!
         Some((timeline, range)) => {
             use std::ops::RangeInclusive;