+++ crates/re_viewer/src/ui/view_text/ui.rs
@@ -29,11 +29,8 @@ pub(crate) fn view_text(
     let time = ctx
         .rec_cfg
         .time_ctrl
-        .time_query()
-        .map_or(state.latest_time, |q| match q {
-            re_data_store::TimeQuery::LatestAt(time) => time,
-            re_data_store::TimeQuery::Range(range) => *range.start(),
-        });
+        .time_i64()
+        .unwrap_or(state.latest_time);
 
     // Did the time cursor move since last time?
     // - If it did, time to autoscroll appropriately.
@@ -96,7 +93,7 @@ fn show_table(
     let item_spacing = ui.spacing().item_spacing;
 
     let current_timeline = *ctx.rec_cfg.time_ctrl.timeline();
-    let current_time = ctx.rec_cfg.time_ctrl.time().map(|tr| tr.floor());
+    let current_time = ctx.rec_cfg.time_ctrl.time_int();
 
     let mut builder = egui_extras::TableBuilder::new(ui)
         .striped(true)